[FIX] sale_timesheet_enterprise: fix the access error while selecting project
Before this commit, when user with the 'timesheets: own documents' and
the 'sales: own documents' access right level try to add the new timesheet
with 'add new line' wizard then it could generate and access on selection of
the project if, he/she does not own the sale order the timesheet is linked to.

So in this commit, use a sudo to access the order's user in compute to
check timesheet has so access or not.

task-2774812

closes odoo/enterprise#27571

X-original-commit: e5fad3363ba744f6f2f50db0d516a24f8d14ade7
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>