[REF] documents_spreadsheet_bundle: revamp data source management
With this commit, we simplify the management of data source by moving
the code of data source to this codebase and clarify the role of the
differents actors:
* DataSources:
Its role is to give an entry point to create, retrieve, load a data source
It should also trigger an event whenever a datasource is updated.
* DataSource
Its role is to manage the fetching of data (and metadata). The data are
contains in a DataSourceModel
* DataSourceModel
Maintain the data of a DataSource

It also contains the following commit of o-spreadsheet:
[REM] data_source

Part of task id 2779455 and 2735400

closes odoo/enterprise#24567

Related: odoo/upgrade#3366
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>