[IMP] {project,sale_timesheet}_enterprise: move profitability section
Before this commit, the mobile view of the profitability section was
displaying from sale_timesheet_enterprise, but in coommunity the
profitability section has been moved from sale_timesheet to project
module.

This commit moves the mobile view of this profitability section shown in
the project update and updates the view too.

task-2710808