[IMP] sale_{project,timesheet}_enterprise: change Sold Section to Sales one
In community version, the Sold section has been replaced by a new
section called Sales. That new section will displayed all the SOLs
related to the project shown in the project update.

This commit replaces the Sold section by the Sales section to have the
new section in the mobile view of the project update view. Moreover, the
this section is moved from `sale_timesheet_enterprise` module to
`sale_project_enterprise` module.

task-2710808