[IMP] l10n_cl_edi: Pull flat discounts from incoming XML
The issue:

The DTE XML sometimes includes flat discounts in the 'Detalle' section
which are indicated using the `DescuentoMonto` tag.

At the moment, these are not taken into account when generating a
vendor bill from an incoming XML file.

(Only percentage discounts, indicated using the `DescuentoPct` tag,
are taken into account).

The solution:

Take the `DescuentoMonto` tag into account as well.

Video explaining the issue and the solution:
https://drive.google.com/file/d/13J3DEx9iBdNrFhqjNQOaeWxCZ1KxkxtG/view?usp=sharing

closes odoo/enterprise#25397

X-original-commit: 133a7c9d2fc56bb848e9690346338a6dae0d8670
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>