[FIX] web_mobile: mobile app crashes when "Add to Home Screen"
In the Android mobile app, in the UserMenu, clicking on the "Add to Home
Screen" results into a crash as the native code receives a full data URL
for the icon instead of only the base64 encoded image.

This commit fixes it by properly trimming the prefix to keep only the
base64 encoded image.

Also, by creating the related test, another issue was detected when a
notification is triggered when the "Add to Home Screen" button is
clicked on the HomeMenu.

task-2766273

closes odoo/enterprise#25320

X-original-commit: 52b87c2ae42782702b158b2cfcc0f8d0d7dc6a76
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>
Signed-off-by: Pierre Paridans (app) <app@odoo.com>