[FIX] account_reports: Fix unfolded 'group_by' in financial report with sum_if_(pos|neg)(_groupby)
In case of 'sum_if_pos'/'sum_if_neg'/'sum_if_pos_groupby'/'sum_if_neg_groupby', the totals of the financial report line was well computed.
However, when unfolding the line, the displayed grouped lines were always computed like a 'sum'.

For example, assume a financial report line with 'sum_if_neg_groupby' as formula considering 2 lines: one +2000 on A, one -10000 on B.
In this case, the total of the report line is correct (-10000). However, when unfolding the line, we get
A: +2000
B: -10000
... instead of only B.

closes odoo/enterprise#25988

Task: 2791942
X-original-commit: 8b17c65a7c86422c4b8e645224a5ff8e0eaacbfa
Signed-off-by: Olivier Colson <oco@odoo.com>
Signed-off-by: Laurent Smet <las@odoo.com>