[FIX] account_winbooks_import: partners sharing bank
Since journal entries and invoices are posted during the import [1], the
partner is required on all the lines before posting.
This wasn't always the case because some partners can be sharing the
same IBAN [2]; in that case the partners were merged and only one reference
was kept.
This commit keeps all the references to partners registered in winbooks,
but assigns an IBAN only on one of the partners if we don't want to
soft-fail the import. Otherwise it will warn the user why the import
failed.

[1] https://github.com/odoo/enterprise/commit/9f368ae949455dd36e94f6e332fbd49587bd45de
[2] https://github.com/odoo/odoo/blob/cffc732e58c14541abcf06bb0d58add224396b7c/odoo/addons/base/models/res_bank.py#L89

closes odoo/enterprise#25449

X-original-commit: 7f98b2bd9fdf275c95a84179075ffcff25e7acc6
Signed-off-by: Quentin De Paoli <qdp@odoo.com>
Signed-off-by: Olivier Dony <odo@odoo.com>
Signed-off-by: William André (wan) <wan@odoo.com>