[IMP] web_gantt: allow multi_group with default_group_by in gantt
Before this commit, the `default_group_by` attribute for the gantt
support only one field. So, the default group by can only be on one
level.

This commit allows giving more than one field to that attribute. So, the
default group by will can be in multi level in the gantt view. To do
that, the attribute can containis many fields separated by a comma.
For instance:

```xml
<gantt default_group_by="project_id,user_id"></gantt>
```

This example will group the records by `project_id > user_id` in the
gantt view.