[IMP] web_gantt: change generic pill progress behavior
In planning, we hatch pills to indicate that a shift is in draft, and we darken the pill to indicate the progress.
The native gantt progress bar hatches the pill to indicate its progress. It is thus incompatible with the current implementation we have in planning.
Nathan (ngr) has agreed to make the darkening of the pill (instead of hatching it) the generic behavior of the native gantt pill progress.

This commit therefore changes the generic behavior of the native gantt pill progress, and makes use of it in planning.

Task-2664116

Part-of: odoo/enterprise#23342