[FIX] documents_spreadsheet: Indeterministic step in tour
One of the tour tests was indeterministically failing on runbot on its
last step as it left a python thread open.
The step in question try to re-open a spreadsheet. During this step, two
rpcs are done:
- "join_spreadsheet_session" which retrieves the data of the spreadsheet
- "dispatch_spreadsheet_message" which send the position of the client
to all other clients.

As the step does not wait for something, the browser is shutdown. It this
shutdown occurs between the two rpcs, the test pass. But if it's occurs
after the two, a python thread is still open, which leads to a test fail.

As the step in question in not relevant, we just remove it.

runbot id: 3500

closes odoo/enterprise#25201

X-original-commit: faca75cb2abe009c811b3d8c749a9056eeecbbec
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>