[IMP] planning, project{_timesheet}_forecast: clean gantt view inheritance
The view `planning.planning_view_gantt` is intherited multiple times to create
views with default groups and change the its js class in different modules.

However, the inheritance tree is a bit confusing and doesn't produce the intended behaviors.

To fix this, this commit cleans the inheritance tree across the modules.

Task-2664116

closes odoo/enterprise#23342

Related: odoo/upgrade#3344
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>