[IMP] hr_contract_salary: give correct salary configurator values
When an addendum to an existing contract is made with a changes of function,
the addendum will take details from the new function
(through the default contract of the new job position).
Previously, the details didn't change and were based on the current
contract details.

taskID 2783059 [Payroll] Addendum to contract doesn't take good values

Part-of: odoo/enterprise#25209