[FIX] account_reports : add comany_id in tax report audit domain
To reproduce
============

- Create two companies
- Create tax in each comany with same value (21% for example)
- Make invoices in each comany with this tax

In tax report with these two companies selected, we notice that the tax 21%
has just the value for the displayed company, but when we go in the audit of this
tax, we notice that the two records of each company is present.

Wanted behavior
===============

We want to see the records of all selected companies both on the report and the audit if
the system parameter `account_tax_report_multi_company` is set.
Otherwise we want to see only the displayed company records on the report and the audit even if
multiple companies are selected.

Purpose
=======

The issue is that while going to the audit the domain doesn't contain a condition on `company_id`.

Specification
=============

To solve the issue, the condition on `company_id` was added to the domain depending on the system
parameter `account_tax_report_multi_company`.

opw-2769251

closes odoo/enterprise#25629

X-original-commit: d327bc2565e7d63830a7bb806fef2aca8dcc86d6
Signed-off-by: Olivier Colson <oco@odoo.com>