[FIX] account_intrastat_expiry: remove grouping on intrastat_id
Intrastat_id is not stored anymore on product in 15.0 (see https://github.com/odoo/enterprise/commit/905c1eca857b57dcbe4ae3cf51421f2a282a14a5 )
The intrastat report has a warning that links to a list of products with problematic intrastat codes configured. This list is grouped by intrastat_id by default. As intrastat_id is not stored in 15.0, the grouping is not possible and it triggers an error message preventing to display the list.
Group by removed in stable.

closes odoo/enterprise#25967

X-original-commit: 6bbbb12f683103dcb9742cf75588126b4d2829b6
Signed-off-by: Olivier Colson <oco@odoo.com>
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>