[IMP] appointment: reorganize answers and description in submission route
In order to avoid code duplication and clarify its content, structure
of submission route is reworked: first a structure is created to centralize
all the answers that the attendee has given in the form in a clear way.
Then, depending on the question type, the description is built and answer
inputs are created adequately.

Also, this method makes sure we do not do excessive computations since we
only go through questions of the appointment type that have been answered.
Also we do not go through all the existing answers in choice questions.

Task-2621327

Part-of: odoo/enterprise#20433