[FIX] sale_planning: remove planned dates into the domain for plan orders
Before this commit, when the domain passed in parameter to the
`action_plan_sale_order` method contained the start_datetime or end_datetime
in its tuples then a traceback can occurs if a bool operator is a leaf.
For instance, if the domain is the following one:
`['&', '|', ('start_datetime, '!=', False), ('end_datetime, '!=', False), ('role_id', '!=', False)]`
then `('start_datetime, '!=', False)` and  `('end_datetime, '!=', False)`
need to be removed because that method used that domain without the
condition on the planned dates to search the shifts unscheduled with the
additional domain found in the `view_domain`. Unfortunately, if those
tuples are removed then the domain will be not valid because a bool
operator will be a leaf.

This commit fixes the issue by replacing the tuple `('start_datetime',
'!=', False)`, for instance, by `('start_datetime', '=', False)` and
also the `(end_datetime, !=, False)`, for instance, by
`('end_datetime', '=', False)` instead of removing those tuples.

task-2742021

Part-of: odoo/enterprise#23824