[IMP] account: select multiple journal groups in multi-companies
In mono-company, journal groups are exclusive, which makes sense.

However, in multi, the user should be able to select one (and only one) journal group per company.

This would for example allow the user to apply the same filter in the displayed journals for each of the companies.

This PR allows for this behaviour, while not changing the way mono-company works.

task-2784137

closes odoo/enterprise#25518

Signed-off-by: Laurent Smet <las@odoo.com>