[IMP] web_gantt,sale_planning: process the actionDomain for plan dialog
Before this commit, when the `actionDomain` contains the planned dates, we
are not able to easily remove the tuples evaluating the planned dates. In
fact, when the user clicks on the magnifying glass, he wants to see the
unschedule shift and not all the shifts or not only the ones schedules.
If the actionDomain is unchanged, then no shifts will be found because
the following domain is added to the actionDomain and given to the plan
dialog:

`['|', ['start_datetime', '=', false], ['end_datetime', '=', false]]`

So, the result of both domains will find no record because the action
domain will search the planned dates for instance and the additional
domain wants only the ones unscheduled.

This commit processes the action domain to replace tuples containing the
planned dates by a `TRUE_LEAF` (= `[1, '=', 1]`) or
`FALSE_LEAF` (`[0, '=', 1]`). If the action domain contains no OR operator
then we could replace all tuples using the planned dates by `TRUE_LEAF`,
because the action domain has to remains valid as before even if we
change a tuple into it.
For instance:

```js
actionDomain = [['role_id', '=', false], ['start_datetime', '!=', false], ['end_datetime',
'!=', false]];
```

The action domain should be satisfied even if we remove the tuples
containing the planned dates. That is, `['role_id', '=', false]`. The
approach to remove those tuples in the action domain has been
experimented but not implemented because of the domains explicit
bool operators. For instance:

```js
actionDomain = [['role_id', '=', false], '&', ['start_datetime', '!=', false],
    ['end_datetime', '!=', false]];
```

If the tuples containing the planned dates are removed, the action
domain will not be valid because the `AND` operator will be a leaf. To solve
this issue, the tuples will be replaced by `TRUE_LEAF` and so the action
domain will be valid.

Another case to solve with this approach is when the tuples to replace
are in the OR operator, for instance:
```js
actionDomain = ['|', ['start_datetime', '!=', false], ['role_id', '!=',
false]]
```

Here, the `['start_datetime', '!=', false]` cannot be replaced by a
`TRUE_LEAF` otherwise the domain will be always satisfied even if the
`role_id` is false. The solution here is to replace the tuple by a
`FALSE_LEAF`. It will mean that the role_id must be set to satisfy the domain
and so the domain works as expected without the planned dates conditions.

Last case study with the same approach is when 2 conditions to replace
has an `OR` operator before them. For instance, let's take this action
domain:

```js
actionDomain = ['|', ['start_datetime', '!=', false], ['end_datetime',
'!=', false]]
```

In that case the tuples to replace cannot be replaced by a `FALSE_LEAF`,
otherwise the domain will always be falsy. To solve this issue, the
approach is to replace the whole expression by a `TRUE_LEAF`. In that
example, the action domain will be equal to the `TRUE_LEAF`. However,
we could functionally assume that the domain in the last example will
never be set on the gantt view since both dates need to be set in order
for te pill to be displayed in the view.

task-2742021

closes odoo/enterprise#23824

Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>
Co-authored-by: William Braeckman <wbr@odoo.com>