[FIX] sale_subscription: hide tokenization checkbox when paying a subscription
Users should not be able to discard the tokenization of their payment if
they pay for a subscription.

This commit introduces a new helper method to better predict when the
tokenization checkbox should be shown or hidden. This also ensures that
the behavior will be the same across all modules.

task-2695201

closes odoo/enterprise#27568

X-original-commit: 9d3713ad16faef6925062efeeb2d5d40e34c73a8
Related: odoo/odoo#91860
Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>