[IMP] industry_fsm(_sale_report): align the footer total in worksheet report
Before this commit, in the field service > worksheet report, the untaxed
amount, taxes, and total labels are aligned to the left, and their values are
aligned to the right end.

So in this commit moves the untaxed amount, taxes, and totals next to their
respective values.

task-2762166

closes odoo/enterprise#25129

Related: odoo/odoo#86103
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>