[FIX] sign: fix navigator not opening wizards for sign items added while signing
Since 15.0, the SignItemNavigator was not correctly opening signature wizards
for sign items added while signing. This occurs because the html structure
between previously added sign items and sign items added while signign are
different. Therefore the SignItemNavigator was triggering the click on the
wrong element. This commit fixes this issue.

task-2783947

closes odoo/enterprise#29317

X-original-commit: b0dc0b28cc58b7614e9171a4869577425f14b20f
Signed-off-by: Arnaud Joset <arj@odoo.com>