[IMP] pos_l10n_se: download saleDetailsReport from closePosPopup
In this commit we make it possible for a user to
download the Sales Report of the current session
by clicking at the "download" icon at the lower
right corner of the closing pos popup.

This is a commonly used feature. So it will
result at a more pleasant UX if we expose
it from the front-end.

The overriding of the funtction `_get_report_values`
in this localisation module is no longer needed
since a more general approach is followed at the base
module (point_of_sale). For this reason we delete it.

task-2755384

closes odoo/enterprise#25910

Related: odoo/odoo#88006
Signed-off-by: Masereel Pierre <pim@odoo.com>