[FIX] documents_spreadsheet: notify snapshot created
When a user created a snapshot, the message `SNAPSHOT_CREATED`
never reached the spreadsheet because the id was missing.
Since the message is missed, the client does not have acknoledged
the last server revision. All following operations are therefore sent
with the wrong server revision id and they are all refused.

closes odoo/enterprise#25435

X-original-commit: 3a96b4779b0bad71e9f3688e19beb5002d63ca6a
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Co-authored-by: Paul Morelle <pmo@odoo.com>