[FIX] helpdesk: fix portal ticket visibility
Expected Behavior
=================
When a helpdesk team privacy is set to "portal", portal users who are followers of a ticket or its team should be able to view that ticket.

Current Behavior
================
Only a portal user set as a partner on the ticket can see it, regardless of the team visibility

This commit replaces the current behavior with the expected one, and also adds tests.

Related: https://github.com/odoo/upgrade/pull/3468
Related: https://github.com/odoo/enterprise/pull/20190

Task-2800363

closes odoo/enterprise#26835

X-original-commit: 6b27fceec3191a59c539df20db5d3b3f6f39f654
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>