[IMP] project_timesheet_forecast(_contract): auto generate timesheet based on contract
Purpose of the commit is to take the contracts working calendars into
account for auto generated timesheet of the planning.

For Ex: An employee as a working time changes during month A (from a full time
to a half time (every morning). A the end of month A, the manager generates the
timesheet for the past month.
When generating the timesheets,
  - All the slots with a date_end prior to the
    working time change should take 8h per day.
  - If there is a slot with a start_date prior to the working time change and
    a end_date after the working time change, it should take 8h before the working time
    change and 4 after.
  - All the slots having a start_date after the working time change should take 4 hours.

Also when employee is archived then employee_id was not set on the planning even
the role_id is set on it so added the employee_id for archived employee.

task-2793011

closes odoo/enterprise#26309

Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>