[IMP] approvals: allow approvers to see the approval requests
In Approvals, if the user is added to an approval request as approver or if
an approval requires the manager to approve but the approver or manager hasn't the
"Approve all requests" security group, he won't be able to see the request to approve.
At the same time, if the user is part of the "Approve all requests" security group,
he has access to all requests made and not only the ones where he is an approver.

This commit adds the ability for a user to see the requests where he is an approver or
the requests of his reports if the approval request requires manager approval.

task-2742881

closes odoo/enterprise#25476

Signed-off-by: Kevin Baptiste <kba@odoo.com>