[FIX] helpdesk: `HelpdeskDashboard`, full height of cells in Safari
Prior this commit, the height of the cell contents in
`o_helpdesk_banner_table` didn't have a full height in Safari, which
created an inconsistency in the table.
This issue is related the a strange behavior, in Safari, when a padding
or a border and the `h-100` class are applied on the same element in
cells.

This commit fixes this issue by removing the button border and
separating the padding and the `h-100` class with an intermediate
element.

task-2754091

closes odoo/enterprise#24154

Signed-off-by: Pierre Paridans (app) <app@odoo.com>