[FIX] voip: phone field _onClick triggered twice
This commit is the counterpart of a community fix
in the web module. The click handler is only
triggered once and a phone call will no longer
raise an error.
A test has been added in the voip module to check
that the handler is only called once.

PR community: odoo/odoo#86928

closes odoo/enterprise#25558

X-original-commit: c4b5a2be44050004ea900c81ff6d660d8c974ed5
Related: odoo/odoo#87196
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>