[FIX] documents_project: fix the access error on clicking of 'create a task'
Before this commit, when user click on the 'create a task' button
leads to an access right error due to creation of a private task
without the user assigned to that task.

So this commit fixes the issue by setting current user as assignee
in that private task.

task-2784885

closes odoo/enterprise#27234

X-original-commit: 1e8987cccabd385a0243807248ba3df25bb209d5
Related: odoo/odoo#91170
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>