[FIX] l10n_mx_edi: retrieve uuid when requesting SAT status of canceled invoice
The issue
At the moment, requesting the SAT status of a cancelled invoice
returns ‘Not Found’. This is because the cancellation process
deletes the cfdi attachment, on which several stored
computed fields depend, notably the Fiscal Folio uuid.
Without the uuid, we can't query the invoice’s status from the SAT.

See this video by @vbe-odoo demonstrating the problem functionally:
https://us02web.zoom.us/rec/play/7kRPfbcM1YixbiezgtHV0C9KAP6AwAmEzf05WtBJO2j3cLMpvSXznOad0qT2u0NoJsisH5ICFLfcxrNN.NjyNJHvBxfK2TNf3?continueMode=true&_x_zm_rtaid=xzBnPaTGTZiTv53zobKqqg.1646836564724.d211a63c07c388cdef9dfb832d780831&_x_zm_rhtaid=552
password: ?wgK1c$N

See this explanation of where the issue occurs in the source code:
https://drive.google.com/file/d/1z8GhsSk6nphmhB46C0wiDHMZjV7_QSZ8/view?usp=sharing

The solution
In stable: Retrieve the uuid from the attachment when querying the SAT
status. (this PR)
In master: Don’t delete the attachment when the invoice is
successfully canceled (odoo/odoo#87142)

Does it work?
Yes. (See this video where I do the testing):
https://drive.google.com/file/d/1hac2dEwU6oFWDcLIC_CvxEIMObozQVY_/view?

Related support tickets
opw-2790491 opw-2716731

closes odoo/enterprise#25673

X-original-commit: f9fd069a8e9872f25e7a16bf1031cd79cdb181e8
Signed-off-by: Josse Colpaert <jco@odoo.com>