[FIX] timesheet_grid: prevent test fail due to holiday generation
Prior to this commit, `test_timesheet_grid_filter_equal_string` was
failing in master since 24th of march 22. This is due to the
generation of a global holiday in the demo data.

This commit changes the date that is freezegun in the past preventing
the demo data to be generated at that time.

closes odoo/enterprise#25565

X-original-commit: 72a2f39c34c3a3b272c66bcb449741558de6e952
Signed-off-by: Xavier <xbo@odoo.com>
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>