[FIX] sale_subscription : access only subscriptions for current company
Steps :
Install Subscription.
Create a 2nd company (say B, and A the first one).
Create a Subscription for Portal in B,
and make sure you already have 3 Subs for him in A.
Log in with Portal in A.
On the home page, note 'Subscription - x' (x the Subs count in A).
Click on it.

Issue :
You see all your Subs, A and B included (more than x Subs).

Fix :
If user has access, count with sudo.

opw-2802829

closes odoo/enterprise#25619

X-original-commit: 08556a215cfba1f007f9a78f802501e5409933f4
Signed-off-by: Arnaud Joset <arj@odoo.com>
Signed-off-by: Onockx Audric (auon) <auon@odoo.com>