[IMP] hr_payroll: fix active notes in multiple lines
In the dashboard of hr_payroll, when there are multiple notes, the tabs could
take more than one line. Since the active tab is differentiated from the others
only with a white border on the bottom, it gets hard to differentiate the
active note if there are multiple lines. This commit fixes this issue using a
grey background-color for the non-active tabs and removing the white border.
Also, newly created notes are now focused automatically so the user can change
their name.

task-2759601