[FIX] project_{forecast,timesheet_forecast}: fix the menu visibility
Before this commit, when disabling 'planning' feature in the settings
of project, which results in disappearance of the schedule menu item
by resource & by role menus due to empty action in menu.

so in this commit fixes the issue by using _uninstall_hook method, to stop the
menu item overriding; which was the reason for the above-mentioned issue.

task-2790017

closes odoo/enterprise#28103

X-original-commit: 7009d87e0f3fd83da11cde1f94d7dc8cc3b6aea7
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>