[REF/FIX] web_enterprise: refactor and fix expiration panel
Previously, when the user's subscription had expired, we would block the
UI. In 14.0 and under, the expiration panel was moved into the blockUI
overlay so that the user could still interact with the expiration panel.
With the wowl refactoring, the ui service doesn't expose an API to do
anything similar, as the desired semantics for such an API are unclear
(in particular: in the case of multiple calls to blockUI with different
content, which should we choose? Should we stack them? Should unrelated
code be allowed to unblock the ui?)

This commit fixes that by introducing a main component that will function
similarly to the blockUI overlay and that contains a different copy of
the expiration panel. Since both expiration panels should share most of
their state, the parts of the logic that are related to managing the
enterprise subscription have been extracted to the enterprise service
(which was renamed to enterprise_subscription service to better match
its purpose). With that, when the subscription is exprired, the home
menu will signal on the bus that the UI should be blocked by the
expiration panel, and when the subscription is almost expired, it will
render the expiration panel within itself as usual.

closes odoo/enterprise#25587

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>