[FIX] l10n_eu_oss_reports: fix tax report with oss
The aim of this commit is to allow amls with other tags than OSS to appear
in the tax report.

Context:
Previously, the oss tax report was supposed to be independent from the
national tax report but the feature evolved and it is now required.
The tags on the amls have to be considered by the tax report but the tax
closing entry shouldn't consider those.

Before this commit:
- amls wearing the OSS tag amongst other tag weren't taken into account to
compute the tax report line of other tag.
- amls wearing the OSS tag weren't taken into account to compute the
tax closing entry.

After this commit:
- amls are taken into account to compute the tax report even if they wear
the oss tag.
- amls wearing the OSS tag aren't taken into account to compute the tax
closing entry.

bugfix ticket: #2796340
support ticket: #2768622, 2768634

closes odoo/enterprise#26303

Enterprise-pr: https://github.com/odoo/enterprise/pull/25591
X-original-commit: 685e031485492e8ac4e234c8d88b1e81bc39bda6
Signed-off-by: Olivier Colson <oco@odoo.com>
Signed-off-by: Brice Bartoletti <bib@odoo.com>