[FIX] l10n_mx_reports: fix unintended edit of countries
Description of the issue
------------------------

The inverse method for l10n_mx_nationality on res.partner
allow to edit the field demonym on res.countries with sudo access

User may enter wrong data and thus modified the res.country and
all the partner linked to the same country. This lead to wrong
info on DIOT report

Solution
--------
If demonym on countries is wrong it should be fixed on the res.country
directly by users that have the access right to modified it

So this inverse method need to be removed.

closes odoo/enterprise#25844

X-original-commit: c71f9560e576cd419d01c695fe4f503b2b964f09
Signed-off-by: Josse Colpaert <jco@odoo.com>