[FIX] web_studio: call super constraint on field names when overriding
web_studio overrides the _name_check constraint on IrModelField
but the parent validation is skipped.
This allows the creation of custom fields that have wrong characters
including the space character, which later causes an error when
'reflect fields' is called and tries to add model_data records with an xml_id
that includes the field name. An sql constraint on xml_id, 'name_nospaces',
causes failure.

A migration script should fix the records that violate the following:
- prefix is `x_`
- no spaces in the field name
Other restrictions in the base model field name are not automatically fixed.

Also, databases with existing records that violate the base constraint
will throw errors when attempting to upgrade the modules on the same version.

closes odoo/enterprise#26519

X-original-commit: acbb99cb3edca31a350a25ae5c50741e0043d17d
Signed-off-by: Christophe Simonis <chs@odoo.com>