[FIX] social_youtube: generate the UTM name based on the Youtube title
Bug
===
When we create a social post for only Youtube, an error is raised
because the name of the UTM source is generated based on the content
(_rec_name). For the social post, this "content" field is "message",
but for Youtube, this field is not used.

Now, we generate the UTM name based on the Youtube title if defined.

Task-2744125

Part-of: odoo/enterprise#25643