[FIX] l10n_nl_intrastat: old transaction code defaults creating traceback
When the default transaction code is being used the CBS export on the
intrastat report, an error occurs. The reason is that the value gathered
from the _fill_missing_values function, which was implemented on the CBS
export function to bring it in line with the content of the report in a
previous fix 3f1c8ce, is an integer, whereas the .join in the export
function expects a string.

closes odoo/enterprise#25675

X-original-commit: 79eded2324dd81699bbbba5aa44af928e2c46ddc
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Daniel Kosky (dako) <dako@odoo.com>