[IMP] documents_spreadsheet_bundle: use position in pivot formulas
The purpose of this task is to support an equivalent of pivot.position
in pivot formulas. It allows to get the nth value of a group by instead
of a concrete value.

For exemple, `=PIVOT(1, "#create_date", 1) will return the value for the
first date we have in the group by.

To distingish the concrete value or the positional value, we use the
character "#" at the beginning of the field name.

Task-id 2802557

closes odoo/enterprise#26125

Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>