[FIX] stock_barcode: Don't show success message on cancel validation
When pressing the "x" button on the immediate picking view when trying to validate a specific picking, the view will no longer behave as if the user validated the picking. Instead, it will now just close the view as if the user had pressed the cancel button.

Task: 2804747
Enterprise PR: https://github.com/odoo/enterprise/pull/25656

closes odoo/enterprise#25819

X-original-commit: a680071e2da1c1198065e7ac706f77a773787f4f
Signed-off-by: Steve Van Essche <svs@odoo.com>