[FIX] planning: fix the archived employee's avatar display in various view
Before this commit, when an employee is archived, the avatar for that archived
employee is not displayed in gantt,list,calendar view of palnning.

so in this commit, pass the employee of the archived resource to display
the avatar of each employee in the planning gantt view also make the avatar
field as compute to fetch the archived employee related to the resource.

task-2774812

X-original-commit: 621ce706d16975886101ce3074b0bad34ef01916
Part-of: odoo/enterprise#27592