[FIX] helpdesk: fix the rating view from helpdesk team
Before this commit, when we have single rating and open the
rating view through the helpdesk team stat button then it will
open the new form view due to the false res_id in action.

So this commit fixes the issue by passing the proper res_id while
opening the form view of rating.

task-2772369

closes odoo/enterprise#27414

X-original-commit: 5acbf94b50ab460455a235b2f6508b7ff25abfb3
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>