[FIX] project_timesheet_forecast_sale: test adaptation
Due to the changes in the community PR, the resource of the employee
is now created directly with the right company (the one provided in create values,
or the environment company, as specified in the default field value).
This means that now, the resource.resource create receives the correct company
and correctly extrapolates the calendar and tz (if not provided).

Previously, when the company/user/calendar/tz were not provided, the resource
tz default used the context/user tz/'UTC'.

Which means that the employees created in the sale_timesheet test common
(including manager_company_B used in the current test) had conflicting
resource and resource calendar tzs, aka 'UTC' for the resource (last default value)
and 'Europe/Brussels' for the resource calendar tz, leading to the previous
result in the test values.

Now that the timezone are the same, the test must be adapted accordingly.

closes odoo/enterprise#25664

Related: odoo/odoo#85601
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>