[IMP] hr_payroll: show first contract date on new employees button click
In the hr dashboard, when you click on the 'new employees' link in the employee
trends section, a list view should appear with the employees that are new.
In this case, it is also interesting to show the first contract date field,
which is optional=hidden. This commit creates a new view that shows this field
in the list view.

task-2759601