[FIX] account_reports: copy of financial reports: proper code replacement
Before v14, formulas on financial reports used to be written in the form 'balance = sum.balance', while in v14, the same formula is expressed by 'sum'.

When copying a financial report, we generate new codes for all its lines, and replace the old ones in all formulas using them within the report. However, this replacement in formulas was made in such a way that it was only compatible with pre-v14 formulas, and wasn't working anymore.

closes odoo/enterprise#25830

X-original-commit: 59b78cf4d8f6356c25ada8ecd1972674acf18877
Signed-off-by: Laurent Smet <las@odoo.com>
Signed-off-by: Olivier Colson <oco@odoo.com>