[FIX] account_reports: fix code replacement in formulas for lines depending on another root line
When copying a financial report, we also copy its lines, and assign new codes to them so that their codes stay unique. We also replace the old codes in all the formulas using them in the report.

Before this fix, the replacement of the codes in the formulas using them was only working if the line using the code was declared under the same root line as the one defining it.

Consider the following simple example:

1) Define a financial report with two lines, as follows:
    - line A, with code A
    - line B, with code B and formula = 'balance = A.balance'

2) Copy the report

==> In the report copy, line A has code A_copy, B has code B_copy, but B's formula is still referencing A instead of A_copy. This is because the code mapping is reinitialized at each iteration of the loop we do on root lines.

closes odoo/enterprise#25875

X-original-commit: 53d06519b6e8bb5a74cf9613972795b22efd698e
Signed-off-by: Laurent Smet <las@odoo.com>
Signed-off-by: Olivier Colson <oco@odoo.com>