[IMP] {website_}appointment: add avatars option and redesign card view
AVATARS

Add avatars_display selection field on the appointment.type model.
The front-end design depends on that parameter, both on small and
large avatars, on different views (see future work).
If the option is 'show', all will appear. Otherwise,
it will be 'hide' and no avatar will be displayed.

It is set as True on demo data in order to display more features.
The option is in appointment to allow customization also when not
having website. It will impact the views of the customer preview
as well in further work.

The field is a stored compute field. If no value is set for website
appointment_types, 'hide' is set as default value. It is set as
'show' for custom ones.

APPOINTMENT TYPE CARD VIEW

Before, the card view was extremely simplified: only an image,
the adress and the name of the appointment. It now uses a more
complete card including:
- The introduction, displayed as well
- A view of avatars of available users, if avatars_display is
'show', is displayed on the card (max 5, (+ number)), or appropriate
icons otherwise.
- Name emphasized and a view availability button.
The cards remain clickable everywhere.
- The title is not red anymore when unpublished

AVATARS ROUTE

In order to avoid having only the images of website_published users
displayed (which is quite a technical parameter), a new route is
used in order to get the avatar of staff_users. This allows bypassing
the access rules (website_published...) It still checks if the given user
is a staff member of the given appointment type.
The avatar (512 by default or given parameter avatar_size) field is used
in order to have the clean letter placeholder if the user has no image set.
A theme-related color is added as background (visible while the image is
not loaded). The selector is set in appointment since it will also be
used in other appointment views.

Task-2496629

Part-of: odoo/enterprise#24744